Move some tests do a different target
